refactor(#3038): move tmux watcher completion gate
